Move examples to top level - Fixes #421
