--- cirq/examples/bell_inequality_test.py
@@ -1,5 +1,0 @@
-from cirq.examples.bell_inequality import main
-
-
-def test_example():
-    main()
