--- cirq/examples/berstein_vazirani_test.py
@@ -1,5 +1,0 @@
-from cirq.examples.bernstein_vazirani import main
-
-
-def test_example():
-    main(None)
