+++ examples/bell_inequality.py
