+++ examples/bell_inequality_test.py
@@ -1,0 +1,5 @@
+from examples.bell_inequality import main
+
+
+def test_example():
+    main()
