+++ examples/bernstein_vazirani.py
