+++ examples/berstein_vazirani_test.py
@@ -1,0 +1,5 @@
+from examples.bernstein_vazirani import main
+
+
+def test_example():
+    main(None)
