+++ examples/supremacy.py
